requirements.txt
snowflake-snowpark-python
snowflake-ml-python
streamlit